:bricks: implementacion de Arquitectura Hexagonal
