+++ server/src/core/adapters/primary/errors/default-application-error-adapter.ts
@@ -1,0 +1,13 @@
+/* eslint-disable @typescript-eslint/strict-boolean-expressions */
+import { ApplicationErrorModel } from '@core/models/errors/application-error-model'
+
+export class DefaultApplicationErrorAdapter extends Error implements ApplicationErrorModel {
+  public statusCode: number = 500
+  public messages: string[] = []
+
+  constructor (message: string) {
+    super(message)
+    this.message = message || this.name
+    this.messages.push(this.message)
+  }
+}
