+++ server/src/core/adapters/primary/errors/request-validation-error-adapter.ts
@@ -1,0 +1,7 @@
+import { HttpStatusMap } from '@core/models/http/http-requets'
+import { DefaultApplicationErrorAdapter } from './default-application-error-adapter'
+
+export class RequestValidationErrorAdapter extends DefaultApplicationErrorAdapter {
+  name = 'Invalid Request'
+  statusCode = HttpStatusMap.badRequest
+}
