+++ server/src/core/frameworks/express/setup-async-error.ts
@@ -1,0 +1,6 @@
+import { expressAsyncErrorsAdapter } from '@core/adapters/primary/express/express-async-error-adapter'
+import { Application } from 'express'
+
+export async function setupAsyncErrors (app: Application): Promise<void> {
+  app.use(expressAsyncErrorsAdapter)
+}
