+++ server/src/core/models/api/api-model.ts
@@ -1,0 +1,6 @@
+import { Router } from 'express'
+
+export interface APIModel {
+  path: string
+  controller: Router
+}
