+++ server/src/core/models/entities/file-entity.ts
@@ -1,0 +1,8 @@
+import { UUID } from '@core/models/types/custom-types'
+
+export interface FileEntity {
+  filename: UUID
+  filesize: number
+  filetype: string
+  binary: Buffer
+}
