+++ server/src/core/models/entities/product-has-model-entity.ts
@@ -1,0 +1,6 @@
+import { UUID } from '@core/models/types/custom-types'
+
+export interface ProductHasModelEntity {
+  productId: UUID
+  modelId: number
+}
