+++ server/src/core/models/entities/product-image-entity.ts
@@ -1,0 +1,6 @@
+import { UUID } from '@core/models/types/custom-types'
+
+export interface ProductImageEntity {
+  productId: UUID
+  fileId: UUID
+}
