+++ server/src/core/models/entities/prototype-product.ts
