+++ server/src/core/models/errors/application-error-model.ts
@@ -1,0 +1,7 @@
+export interface ApplicationErrorModel {
+  statusCode: number
+  message: string
+  messages?: string[]
+  name?: string
+  stack?: Error['stack']
+}
