+++ server/src/core/models/middlewares/express-middleware-model.ts
@@ -1,0 +1,5 @@
+import { HttpRequestModel } from '../http/http-response'
+
+export interface MiddlewareExpressModel extends HttpRequestModel {
+  method?: string
+}
