+++ server/src/core/models/types/custom-types.ts
