+++ server/src/core/ports/input/middlewares/middleware-input-port.ts
@@ -1,0 +1,5 @@
+import { MiddlewareExpressModel } from '@core/models/middlewares/express-middleware-model'
+
+export interface MiddlewareExpressInputPort {
+  handleRequest: (request: MiddlewareExpressModel) => Promise<void> | never
+}
