+++ server/src/core/ports/output/orm/db-output-port.ts
@@ -1,0 +1,4 @@
+export interface DatabaseOutputPort {
+  connect: () => Promise<void>
+  syncModels?: (alter: boolean) => Promise<void>
+}
