--- server/src/entities/file-entity.ts
@@ -1,8 +1,0 @@
-import { UUID } from '@shared/custom-types'
-
-export interface FileEntity {
-  filename: UUID
-  filesize: number
-  filetype: string
-  binary: Buffer
-}
