--- server/src/entities/product-has-model-entity.ts
@@ -1,6 +1,0 @@
-import { UUID } from '@shared/custom-types'
-
-export interface ProductHasModelEntity {
-  productId: UUID
-  modelId: number
-}
