--- server/src/entities/product-image-entity.ts
@@ -1,6 +1,0 @@
-import { UUID } from '@shared/custom-types'
-
-export interface ProductImageEntity {
-  productId: UUID
-  fileId: UUID
-}
