+++ server/src/server.ts
@@ -2,24 +2,24 @@ import 'reflect-metadata'
 import 'source-map-support/register'
 import 'module-alias/register'
 
-import express from 'express'
-import loaders from './loaders'
-import api from './routes'
-import sequelizeInstance from '@shared/database-config'
+import { ExpressHttpServerAdapter } from '@core/adapters/primary/express/express-server-adapter'
+import { SequelizeAdapter } from '@core/adapters/secondary/orm/sequelize/sequelize-adapter'
+import constants from '@shared/constants'
 
-void startServer()
+void main()
 
-async function startServer (): Promise<void> {
-  const app = express()
+async function main (): Promise<void> {
+  const db = new SequelizeAdapter()
 
-  try {
-    await loaders.init({
-      expressApp: app,
-      expressRoutes: api(),
-      sequelize: sequelizeInstance
-    })
-    app.listen(4500, () => console.log('Server is running on port: 4500'))
-  } catch (error) {
-    console.log(error)
-  }
+  db.connect()
+    .then(() => console.log('database connected successfully'))
+    .catch(error => console.log(error))
+
+  db.syncModels(constants.SYNC_DATABASE)
+    .then(() => console.log('synchronized database models'))
+    .catch(error => console.log(error))
+
+  const httpServer = new ExpressHttpServerAdapter(constants.SERVER_PORT)
+
+  await httpServer.start()
 }
