+++ server/tsconfig.json
@@ -31,11 +31,9 @@
       "baseUrl": "./src",                                  /* Specify the base directory to resolve non-relative module names. */
       "paths": {
         "*": ["types/*"],
-        "@loaders/*": ["loaders/*"],
-        "@routes/*": ["routes/*"],
+        "@core/*": ["core/*"],
         "@shared/*": ["shared/*"],
-        "@entities/*": ["entities/*"],
-        "@models/*": ["models/*"],
+        "@products/*": ["products/*"]
       },                                          /* Specify a set of entries that re-map imports to additional lookup locations. */
       // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
       // "typeRoots": ["./node_modules/@types", "./src/types"],                                  /* Specify multiple folders that act like './node_modules/@types'. */
